periodic table move to element.base module
