+++ pyabc/crystal/structure.py
@@ -4,37 +4,7 @@ from itertools import product
 import spglib
 
 from pyabc.utils.math import closest_pair
-
-"""
-periodic_table_dict
-Thanks for Yanan Wu. Email:
-"""
-periodic_table_dict = {'Vacc': 0,
-                       'H': 1, 'He': 2,
-                       'Li': 3, 'Be': 4, 'B': 5, 'C': 6, 'N': 7, 'O': 8,
-                       'F': 9, 'Ne': 10,'Na': 11, 'Mg': 12, 'Al': 13, 'Si': 14, 'P': 15, 'S': 16, 'Cl': 17, 'Ar': 18,
-                       'K': 19, 'Ca': 20, 'Sc': 21, 'Ti': 22, 'V': 23, 'Cr': 24, 'Mn': 25, 'Fe': 26, 'Co': 27, 'Ni': 28, 'Cu': 29, 'Zn': 30, 'Ga': 31, 'Ge': 32, 'As': 33, 'Se': 34, 'Br': 35, 'Kr': 36,
-                       'Rb': 37, 'Sr': 38, 'Y': 39, 'Zr': 40, 'Nb': 41, 'Mo': 42, 'Tc': 43, 'Ru': 44, 'Rh': 45, 'Pd': 46, 'Ag': 47, 'Cd': 48, 'In': 49, 'Sn': 50, 'Sb': 51, 'Te': 52, 'I': 53, 'Xe': 54,
-                       'Cs': 55, 'Ba': 56, 'Hf': 72, 'Ta': 73, 'W': 74, 'Re': 75, 'Os': 76, 'Ir': 77, 'Pt': 78, 'Au': 79, 'Hg': 80, 'Tl': 81, 'Pb': 82, 'Bi': 83, 'Po': 84, 'At': 85, 'Rn': 86,
-                       'Fr': 87, 'Ra': 88}
-
-
-def get_symbol(atom):
-    """
-    get_symbol return symbol of atomic number
-
-    parameter:
-
-    atom: int, atomic number.
-    """
-    for key, value in periodic_table_dict.items():
-        if atom == value:
-            return str(key)
-    return "NaN_x"
-
-
-def symbol2number(symbol):
-    return periodic_table_dict[symbol]
+from pyabc.element.base import *
 
 
 def car_to_frac(lattice, car_vec):
